rename rsvp block component folders
