+++ plugins/tickets/src/modules/blocks/rsvp/action-buttons/attendees-action-button/template.js
