+++ plugins/tickets/src/modules/blocks/rsvp/action-buttons/index.js
