+++ plugins/tickets/src/modules/blocks/rsvp/action-buttons/settings-action-button/container.js
