+++ plugins/tickets/src/modules/blocks/rsvp/advanced-options/style.pcss
