+++ plugins/tickets/src/modules/blocks/rsvp/attendee-registration/template.js
