+++ plugins/tickets/src/modules/blocks/rsvp/counters/template.js
