+++ plugins/tickets/src/modules/blocks/rsvp/dashboard/container.js
