+++ plugins/tickets/src/modules/blocks/rsvp/duration-label/container.js
