+++ plugins/tickets/src/modules/blocks/rsvp/duration-label/style.pcss
