+++ plugins/tickets/src/modules/blocks/rsvp/duration-label/template.js
