+++ plugins/tickets/src/modules/blocks/rsvp/duration-picker/style.pcss
