+++ plugins/tickets/src/modules/blocks/rsvp/duration-picker/template.js
