+++ plugins/tickets/src/modules/blocks/rsvp/header-image/container.js
