+++ plugins/tickets/src/modules/blocks/rsvp/header-image/template.js
