+++ plugins/tickets/src/modules/blocks/rsvp/settings-dashboard/container.js
