+++ plugins/tickets/src/modules/blocks/rsvp/template.js
@@ -8,8 +8,8 @@ import classNames from 'classnames';
 /**
  * Internal dependencies
  */
-import RSVPContainer from './rsvp-container/container';
-import RSVPDashboard from './rsvp-dashboard/container';
+import RSVPContainer from './container/container';
+import RSVPDashboard from './dashboard/container';
 import './style.pcss';
 
 const RSVP = ( { isSelected } ) => (
